Use precision of timestamps in queries for search tolerance
Adds the date_time_precision dependency which shows the precision of a
DateTime object that we parse from the search query.  When applying a
fuzz/tolerance to the query against a timestamp field, use the precision
to search by the most logical amount.

Fixes the case when searching for "> 2014-05-18 11:45:30" would search
for records after 12:45:30 due to a one hour tolerance being
automatically applied despite the precision of the timestamp in the
original query.