Filter URL parameters from ActionController::Parameters in sort helper
On Rails 5, passing `params` directly into `url_for` results in:

    ActionView::Template::Error (Attempting to generate a URL from
    non-sanitized request parameters! An attacker can inject malicious
    data into the generated URL, such as changing the host. Whitelist
    and sanitize passed parameters to be secure.):

The url_options/params is now parameterised so the user can pass a
filtered list of current URL options to the helper. If the default
`params` is used or an AC::Parameters object is passed, it's converted
to a hash of filtered parameters.