Don't eager load data on :has_many relations
When generating a query to search on a column in a related table that
has a has_many relationship, scoped search uses an inner query to get
the ids of the records to return (using `primary_key IN (SELECT...)`)
when the operator is a comparison operator (eq, ne, gt, etc.).

Therefore, the related table doesn't need to be joined for the query to
return the correct result, reducing the complexity of the query and the
size of the results returned in these cases.