Add an "onClear" option to scopedSearch
Scoped-search's clear function sets an empty value by default.
However, there are cases where one would like to customize that behavior.
For example, set a different clear value (e.g. one space).

In this patch, I introduce a way to do that by setting the onClear option.

In the following example, we confirm that the user wants to clean the
search box before cleaning it:

$(input).scopedSearch({'onClear': function(event) {
  confirm("are you sure?") ? $(input).value('') : null
});

Signed-off-by: Boaz Shuster <ripcurld.github@gmail.com>