Improve search queries based on datetime
The current implementation causes issues such as:

1. When searching for timestamps greater than the given value and
minutes and seconds are omitted the value, it returns only timestamps
greater than value + 1 hour.

2. When searching for timestamps greater than the given value and only
month is provided as the smallest time unit, it excludes only timestamps
from the first day of the month in the results but includes rest of
month. Based on the logic implemented here this should not happen but
instead the whole month should be excluded from the search results, i.e.
'> January 2024' should exclude all January timestamps.

Same problems occur when searching with `<=`.

This PR aims to fix these issues.